#939 Add `@IgnoreProperty` with standard introspector to add API level attribute ignore marker
