+++ hartshorn-config/hartshorn-config-jackson/src/main/java/org/dockbox/hartshorn/config/jackson/introspect/JacksonIgnoreAnnotationIntrospector.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2019-2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.dockbox.hartshorn.config.jackson.introspect;
+
+import com.fasterxml.jackson.databind.introspect.AnnotatedMember;
+import com.fasterxml.jackson.databind.introspect.JacksonAnnotationIntrospector;
+
+import org.dockbox.hartshorn.config.annotations.IgnoreProperty;
+import org.dockbox.hartshorn.util.introspect.Introspector;
+
+public class JacksonIgnoreAnnotationIntrospector extends JacksonAnnotationIntrospector {
+
+    private final Introspector introspector;
+
+    public JacksonIgnoreAnnotationIntrospector(final Introspector introspector) {
+        this.introspector = introspector;
+    }
+
+    @Override
+    public boolean hasIgnoreMarker(final AnnotatedMember member) {
+        return this.introspector.introspect(member.getAnnotated()).has(IgnoreProperty.class);
+    }
+}
