+++ hartshorn-config/hartshorn-config-jackson/src/main/java/org/dockbox/hartshorn/config/jackson/mapping/JsonDataMapper.java
@@ -14,22 +14,24 @@
  * limitations under the License.
  */
 
-package org.dockbox.hartshorn.config.jackson;
+package org.dockbox.hartshorn.config.jackson.mapping;
 
 import com.fasterxml.jackson.databind.cfg.MapperBuilder;
-import com.fasterxml.jackson.dataformat.xml.XmlMapper;
+import com.fasterxml.jackson.databind.json.JsonMapper;
 
 import org.dockbox.hartshorn.config.FileFormat;
 import org.dockbox.hartshorn.config.FileFormats;
+import org.dockbox.hartshorn.config.jackson.JacksonDataMapper;
 
-public class XmlDataMapper implements JacksonDataMapper {
+public class JsonDataMapper implements JacksonDataMapper {
+
     @Override
     public FileFormat fileFormat() {
-        return FileFormats.XML;
+        return FileFormats.JSON;
     }
 
     @Override
     public MapperBuilder<?, ?> get() {
-        return XmlMapper.builder();
+        return JsonMapper.builder();
     }
 }
