+++ hartshorn-config/hartshorn-config-jackson/src/main/java/org/dockbox/hartshorn/config/jackson/mapping/XmlDataMapper.java
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package org.dockbox.hartshorn.config.jackson;
+package org.dockbox.hartshorn.config.jackson.mapping;
 
 import com.fasterxml.jackson.databind.cfg.MapperBuilder;
-import com.fasterxml.jackson.dataformat.toml.TomlMapper;
+import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 
 import org.dockbox.hartshorn.config.FileFormat;
 import org.dockbox.hartshorn.config.FileFormats;
+import org.dockbox.hartshorn.config.jackson.JacksonDataMapper;
 
-public class TomlDataMapper implements JacksonDataMapper {
+public class XmlDataMapper implements JacksonDataMapper {
     @Override
     public FileFormat fileFormat() {
-        return FileFormats.TOML;
+        return FileFormats.XML;
     }
 
     @Override
     public MapperBuilder<?, ?> get() {
-        return TomlMapper.builder();
+        return XmlMapper.builder();
     }
 }
